requirements.txt
flask
requests
python-telegram-bot==13.15
beautifulsoup4
